# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tagg@d API",
    "version": "1.0.0",
    "description": "API documentation for the Tagg@d backend — a crypto payment platform for Africa."
  },
  "servers": [
    {
      "url": "http://localhost:5002/api/v2",
      "description": "Current version (v2)"
    },
    {
      "url": "http://localhost:5002/api/v1",
      "description": "Deprecated version (v1)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Development Server (unversioned root)"
    },
    {
      "url": "https://taggedpay.xyz/api/v2",
      "description": "Production (v2)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
...
  },
  "paths": {
...
    "/api/balances/all": {
      "get": {
        "summary": "Get all balances (admin)",
        "tags": [
          "Balances"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all balances"
          }
        }
      }
    },
...
    "/api/send-to-tag": {
      "post": {
        "summary": "Send funds to a @tag",
        "tags": [
          "General"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sender_tag",
                  "receiver_tag",
                  "amount"
                ],
                "properties": {
                  "sender_tag": {
                    "type": "string",
                    "example": "alice"
                  },
                  "receiver_tag": {
                    "type": "string",
                    "example": "bob"
                  },
                  "amount": {
                    "type": "number",
                    "example": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer initiated"
          }
        }
      }
    },
    "/api/send-to-wallet": {
      "post": {
        "summary": "Send funds to a wallet address",
        "tags": [
          "General"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sender_tag",
                  "receiver_address",
                  "amount"
                ],
                "properties": {
                  "sender_tag": {
                    "type": "string"
                  },
                  "receiver_address": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer initiated"
          }
        }
      }
    },
...
    "/api/wallets": {
      "get": {
        "summary": "Get the authenticated user's wallets",
        "description": "Returns all wallets and balances associated with the authenticated user across all supported chains.",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User wallets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "chain": {
                            "type": "string",
                            "example": "stellar"
                          },
                          "address": {
                            "type": "string",
                            "example": "GABCDXYZ1234567890..."
                          },
                          "balances": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "token": {
                                  "type": "string",
                                  "example": "USDC"
                                },
                                "amount": {
                                  "type": "number",
                                  "example": 500
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/wallets/send-to-tag": {
      "post": {
        "summary": "Send funds to another user via @tag",
        "description": "Transfer tokens to another registered user's @tag on any supported chain (base, lisk, flow, u2u, starknet). For Flow, uses Flow EVM (0x + 16 hex chars).\n",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiver_tag",
                  "amount",
                  "balance_id"
                ],
                "properties": {
                  "receiver_tag": {
                    "type": "string",
                    "example": "alice",
                    "description": "Recipient's @tag (3-20 alphanumeric/underscore)"
                  },
                  "amount": {
                    "type": "number",
                    "example": 1.5,
                    "description": "Amount to send (positive, up to 18 decimal places)"
                  },
                  "balance_id": {
                    "type": "integer",
                    "example": 42,
                    "description": "Sender's balance record ID (determines chain + token)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transfer initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "example": "success"
                    },
                    "txHash": {
                      "type": "string",
                      "example": "0xabc123..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or missing fields"
          },
          "422": {
            "description": "Insufficient balance or transfer failed"
          },
          "429": {
            "description": "Another transaction already in progress"
          }
        }
      }
    },
    "/api/wallets/send-to-wallet": {
      "post": {
        "summary": "Send funds to an external blockchain address",
        "description": "Withdraw tokens to an external wallet address on any supported chain. Flow EVM addresses must be 0x followed by exactly 16 hex characters (e.g. 0x1234567890abcdef). Requires 2FA.\n",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "receiver_address",
                  "amount",
                  "balance_id"
                ],
                "properties": {
                  "receiver_address": {
                    "type": "string",
                    "example": "0x1234567890abcdef",
                    "description": "Destination blockchain address. Format depends on chain: Flow EVM: 0x + 16 hex chars, EVM chains: 0x + 40 hex chars, Starknet: 0x + up to 64 hex chars\n"
                  },
                  "amount": {
                    "type": "number",
                    "example": 0.5
                  },
                  "balance_id": {
                    "type": "integer",
                    "example": 42
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdrawal initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "example": "success"
                    },
                    "txHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid address format or missing fields"
          },
          "422": {
            "description": "Insufficient balance or transfer failed"
          }
        }
      }
    },
    "/api/wallets/{id}": {
      "get": {
        "summary": "Get a specific wallet by ID",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Wallet ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "chain": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "balances": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Wallet not found"
          }
        }
      },
      "put": {
        "summary": "Update wallet settings",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "example": "My Stellar Wallet"
                  },
                  "is_active": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet updated"
          },
          "404": {
            "description": "Wallet not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a wallet",
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet deleted"
          },
          "404": {
            "description": "Wallet not found"
          }
        }
      }
    },
...
  },
...
}